% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_clean_data.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Function to clean the LOCATION_NAME field in the NOAA
earthquake dataset}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{NOAA signification earthquake dataset}
}
\value{
data.frame with cleaned LOCATION_NAME field
}
\description{
Function to clean the LOCATION_NAME field in the NOAA
earthquake dataset
}
\examples{
\dontrun{eq_location_clean(data)}
}
